feat: Increase rows in form text areas
This commit increases the number of rows in the form text areas for the `descr` attribute in multiple views. The rows are increased from 6 to 10 to provide a larger input area for users to enter descriptions.

The affected views are:
- `books/_form.html.erb`
- `families/_form.html.erb`
- `names/_form.html.erb`
- `orbs/_form.html.erb`
- `races/_form.html.erb`
- `ship_sizes/_form.html.erb`
- `ships/_form.html.erb`

This change improves the user experience by allowing for more visible content in the text areas.